+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/download/Storage.java
@@ -25,7 +25,7 @@ import lombok.SneakyThrows;
 import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import org.icgc.dcc.pcawg.client.core.model.metadata.FileMetaData;
+import org.icgc.dcc.pcawg.client.model.metadata.FileMetaData;
 
 import java.io.File;
 import java.io.IOException;
