+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/download/fetcher/Fetcher.java
@@ -1,6 +1,6 @@
 
 package org.icgc.dcc.pcawg.client.download.fetcher;
-import org.icgc.dcc.pcawg.client.core.model.metadata.FileMetaDataContext;
+import org.icgc.dcc.pcawg.client.model.metadata.FileMetaDataContext;
 
 
 import java.io.IOException;
