+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/download/fetcher/decorators/MaxFileSizeFetcherDecorator.java
@@ -4,7 +4,7 @@ import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.val;
-import org.icgc.dcc.pcawg.client.core.model.metadata.FileMetaDataContext;
+import org.icgc.dcc.pcawg.client.model.metadata.FileMetaDataContext;
 import org.icgc.dcc.pcawg.client.download.fetcher.Fetcher;
 
 import static lombok.AccessLevel.PRIVATE;
