+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/download/fetcher/impl/NumDonorsFetcher.java
@@ -2,13 +2,13 @@
 package org.icgc.dcc.pcawg.client.download.fetcher.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.icgc.dcc.pcawg.client.core.model.metadata.FileMetaDataContext;
+import org.icgc.dcc.pcawg.client.model.metadata.FileMetaDataContext;
 import org.icgc.dcc.pcawg.client.download.fetcher.Fetcher;
 import org.icgc.dcc.pcawg.client.download.Portal;
 
 import java.io.IOException;
 
-import static org.icgc.dcc.pcawg.client.core.model.metadata.FileMetaDataContext.buildFileMetaDataContext;
+import static org.icgc.dcc.pcawg.client.model.metadata.FileMetaDataContext.buildFileMetaDataContext;
 
 @RequiredArgsConstructor
 public class NumDonorsFetcher implements Fetcher {
