+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/model/ssm/Common.java
@@ -1,4 +1,4 @@
-package org.icgc.dcc.pcawg.client.core;
+package org.icgc.dcc.pcawg.client.model.ssm;
 
 public interface Common {
   String getAnalysisId();
