+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/model/ssm/primary/AbstractSSMPrimaryBase.java
@@ -1,4 +1,7 @@
-package org.icgc.dcc.pcawg.client.core;
+package org.icgc.dcc.pcawg.client.model.ssm.primary;
+
+import org.icgc.dcc.pcawg.client.core.ProjectMetadataDAO;
+import org.icgc.dcc.pcawg.client.model.ssm.AbstractCommon;
 
 /**
  * Common place for
