+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/model/ssm/primary/SSMPrimary.java
@@ -1,4 +1,6 @@
-package org.icgc.dcc.pcawg.client.core;
+package org.icgc.dcc.pcawg.client.model.ssm.primary;
+
+import org.icgc.dcc.pcawg.client.model.ssm.Common;
 
 public interface SSMPrimary extends Common {
